Made settings and rules screens more consistent with the rest of the app
